% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-methods.R
\name{anova.longmi}
\alias{anova.longmi}
\title{Model Comparison Method for an Object of Class \code{longmi}}
\usage{
\method{anova}{longmi}(object, ...)
}
\arguments{
\item{object}{Object of class \code{longmi}.}

\item{...}{Additional arguments to pass to \code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}}.}
}
\value{
Returns a data frame of chi-square difference test results.
}
\description{
Model Comparison Method for an Object of Class \code{longmi}
}
\examples{
data("osbornesudick1972", package = "longMI")
mi <- Invariance(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
anova(mi)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
