% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-invariance.R
\name{Invariance}
\alias{Invariance}
\title{Test Longitudinal Measurement Invariance}
\usage{
Invariance(
  data,
  time_points,
  factor_loadings,
  covariances = FALSE,
  model_add_configural = NULL,
  model_add_weak = NULL,
  model_add_strong = NULL,
  model_add_strict = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe.
The function assumes that the data is in the wide format
and the variables are named as follows:
\code{paste0("y", time_point, "_", item_number)}.
For example, for the item 1 from the first time point,
the variable name should be \code{y1_1}.}

\item{time_points}{Numeric vector of discrete time points.}

\item{factor_loadings}{List with length equal to the number of factors.
Each element of the list is the item number of items
for the specific factor.}

\item{covariances}{Logical.
If \code{covariance = TRUE},
model the covariances of the measurement error.}

\item{model_add_configural}{Additional specification
added to the lavaan model syntax
for the configural invariance model.}

\item{model_add_weak}{Additional specification
added to the lavaan model syntax
for the weak invariance model.}

\item{model_add_strong}{Additional specification
added to the lavaan model syntax
for the strong invariance model.}

\item{model_add_strict}{Additional specification
added to the lavaan model syntax
for the strict invariance model.}

\item{...}{Additional arguments to pass to \code{\link[lavaan:cfa]{lavaan::cfa()}}.}
}
\value{
Returns an object of class \code{longmi} which is
a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{List of function arguments.}
\item{fit}{Fitted models.}
\item{fun}{Function used ("Invariance").}
}
}
\description{
Test Longitudinal Measurement Invariance
}
\examples{
data("osbornesudick1972", package = "longMI")
mi <- Invariance(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
print(mi)
summary(mi)
}
\seealso{
Other Longitudinal Measurement Invariance Functions: 
\code{\link{Comparison}()},
\code{\link{Configural}()},
\code{\link{Strict}()},
\code{\link{Strong}()},
\code{\link{Weak}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Longitudinal Measurement Invariance Functions}
\keyword{invariance}
\keyword{manSASInvariance}
