% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-package.R
\docType{package}
\name{longMI-package}
\alias{longMI-package}
\alias{_PACKAGE}
\title{longMI: Longitudinal Measurement Invariance}
\description{
Fits longitudinal measurement invariance models using the 'lavaan' package. For a thorough exposition of testing measurement invariance, see Millsap (2011) \doi{10.4324/9780203821961}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/longMI}
  \item \url{https://ijapesigan.github.io/longMI/}
  \item Report bugs at \url{https://github.com/ijapesigan/longMI/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
