% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets-osbornesudick1972.R
\docType{data}
\name{osbornesudick1972}
\alias{osbornesudick1972}
\title{Wechsler Intelligence Scale for Children Data from Osborne and Sudick (1972)}
\format{
The data set has the following variables:
\describe{
\item{id}{ID}
\item{y1_1}{Time 1 Information}
\item{y1_2}{Time 1 Comprehension}
\item{y1_3}{Time 1 Similarities}
\item{y1_4}{Time 1 Vocabulary}
\item{y6_1}{Time 6 Information}
\item{y6_2}{Time 6 Comprehension}
\item{y6_3}{Time 6 Similarities}
\item{y6_4}{Time 6 Vocabulary}
}
}
\usage{
osbornesudick1972
}
\description{
Wechsler Intelligence Scale for Children Data from Osborne and Sudick (1972)
}
\references{
Osborne, R. T., & Suddick, D. E. (1972).
A longitudinal investigation of the intellectual differentiation hypothesis.
\emph{The Journal of Genetic Psychology:}
\emph{Research and Theory on Human Development},
\emph{121}(1), 83–89.
\doi{10.1080/00221325.1972.10533131}.
}
\keyword{data}
