% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-methods.R
\name{summary.longmi}
\alias{summary.longmi}
\title{Summary Method for an Object of Class \code{longmi}}
\usage{
\method{summary}{longmi}(object, ...)
}
\arguments{
\item{...}{additional arguments to pass to the summary function
in \code{lavaan}}

\item{x}{Object of class \code{longmi}
that is, the output of the \code{\link[=Invariance]{Invariance()}}
or the \code{\link[=Comparison]{Comparison()}} functions.}
}
\value{
Returns a list of
the summary of the fitted models.
}
\description{
Summary Method for an Object of Class \code{longmi}
}
\examples{
data("osbornesudick1972", package = "longMI")
mi <- Invariance(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
summary(mi)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
