% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-comparison.R
\name{Comparison}
\alias{Comparison}
\title{Compare Measurement Invariance Models}
\usage{
Comparison(configural = NULL, weak = NULL, strong = NULL, strict = NULL, ...)
}
\arguments{
\item{configural}{Fitted configural invariance model.}

\item{weak}{Fitted weak invariance model.}

\item{strong}{Fitted strong invariance model.}

\item{strict}{Fitted strict invariance model.}

\item{...}{Additional arguments to pass to \code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}}.}
}
\value{
Returns an object of class \code{longmi} which is
a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{List of function arguments.}
\item{fit}{Fitted models.}
\item{measures}{Fit measures.}
\item{fun}{Function used ("Comparison").}
}
}
\description{
Compare Measurement Invariance Models
}
\examples{
data("osbornesudick1972", package = "longMI")
configural_fit <- Configural(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
weak_fit <- Weak(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
strong_fit <- Strong(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
strict_fit <- Strict(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
mi <- Comparison(
  configural = configural_fit,
  weak = weak_fit,
  strong = strong_fit,
  strict = strict_fit
)
print(mi)
summary(mi)
}
\seealso{
Other Longitudinal Measurement Invariance Functions: 
\code{\link{Configural}()},
\code{\link{Invariance}()},
\code{\link{Strict}()},
\code{\link{Strong}()},
\code{\link{Weak}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Longitudinal Measurement Invariance Functions}
\keyword{comparison}
\keyword{manSASInvariance}
