% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longMI-methods.R
\name{print.longmi}
\alias{print.longmi}
\title{Print Method for an Object of Class \code{longmi}}
\usage{
\method{print}{longmi}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{longmi}.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
selected fit measures.
}
\description{
Print Method for an Object of Class \code{longmi}
}
\examples{
data("osbornesudick1972", package = "longMI")
mi <- Invariance(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
print(mi)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
